% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{calculate_omega_constraint}
\alias{calculate_omega_constraint}
\title{function that computes the overlap of two feasibility domains}
\usage{
calculate_omega_constraint(A, B, raw = FALSE, nsamples)
}
\arguments{
\item{A}{one interaction matrix}

\item{B}{the constraint matrix}

\item{raw}{TRUE: raw omega, FALSE: normalized omega}

\item{nsamples}{number of sampled points}
}
\value{
the normalize feasibility of the intersection region
}
\description{
function that computes the overlap of two feasibility domains
}
