% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{.near}
\alias{.near}
\title{Internal helper: approximate equality check (replaces dplyr::near)}
\usage{
.near(x, y, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{tol}{tolerance}
}
\value{
logical vector
}
\description{
Internal helper: approximate equality check (replaces dplyr::near)
}
\keyword{internal}
