% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{generate_span_vectors}
\alias{generate_span_vectors}
\title{function that normalizes the spanning vectors of the feasibility domain in the L2 norm}
\usage{
generate_span_vectors(alpha)
}
\arguments{
\item{alpha}{interaction matrix}
}
\value{
the normalized spanning vectors
}
\description{
function that normalizes the spanning vectors of the feasibility domain in the L2 norm
}
