% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{inside_vertex_detection}
\alias{inside_vertex_detection}
\title{function that computes all the extreme points that belong to original vertexes}
\usage{
inside_vertex_detection(A, B)
}
\arguments{
\item{A}{one interaction matrix}

\item{B}{another interaction matrix}
}
\value{
all the extreme points that belong to original vertexes
}
\description{
function that computes all the extreme points that belong to original vertexes
}
