% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{interaction_matrix_random}
\alias{interaction_matrix_random}
\title{function that generates a random matrix (as in May, Nature (1972))}
\usage{
interaction_matrix_random(num, stren, conne)
}
\arguments{
\item{num}{number of species}

\item{stren}{standard deviation of interaction strength}

\item{conne}{connectance of the interaction matrix}
}
\value{
the generated random matrix
}
\description{
function that generates a random matrix (as in May, Nature (1972))
}
