% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{intersection_vertex_detection}
\alias{intersection_vertex_detection}
\title{function that computes all the extreme points generated that are generated by the intersections of the cones}
\usage{
intersection_vertex_detection(S, M)
}
\arguments{
\item{S}{one interaction matrix}

\item{M}{another interaction matrix}
}
\value{
all the extreme points that are generated by the intersections of the cones
}
\description{
function that computes all the extreme points generated that are generated by the intersections of the cones
}
