% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{normalize}
\alias{normalize}
\title{function that normalizes a vector in the L2 norm}
\usage{
normalize(a, norm = 2)
}
\arguments{
\item{a}{= the original vector}
}
\value{
the normalized vector
}
\description{
function that normalizes a vector in the L2 norm
}
