% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{vertex_detection}
\alias{vertex_detection}
\title{function that computes all the extreme points}
\usage{
vertex_detection(A, B)
}
\arguments{
\item{A}{one interaction matrix}

\item{B}{another interaction matrix}
}
\value{
all the extreme points that generate the intersection region
}
\description{
function that computes all the extreme points
}
