% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{calculate_omega}
\alias{calculate_omega}
\title{function that computes the normalized feasibility from an interaction matrix}
\usage{
calculate_omega(vertex)
}
\arguments{
\item{vertex}{all the vertexes of the feasibility domain}
}
\value{
the normalized feasibility
}
\description{
function that computes the normalized feasibility from an interaction matrix
}
